requirements.txt
# Core
streamlit>=1.40.0
python-dotenv>=1.0.0

# Firebase
firebase-admin>=6.5.0

# Google Fit
google-auth>=2.28.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.120.0

# AI / ML
openai>=1.12.0
scikit-learn>=1.4.0
numpy>=1.26.0
pandas>=2.2.0

# Visualization
plotly>=5.18.0

# Utils
requests>=2.31.0
Pillow>=10.2.0
